+++ lib/Models/MapInteractionMode.ts
@@ -15,6 +15,7 @@ interface Options {
   buttonText?: string;
   uiMode?: UIMode; // diff tool hack for now
   onEnable?: (viewState: ViewState) => void;
+  invisible?: boolean;
 }
 
 /**
@@ -25,6 +26,8 @@ export default class MapInteractionMode {
 
   readonly buttonText: string;
   readonly uiMode: UIMode;
+
+  readonly invisible: boolean;
 
   @observable
   customUi: (() => any) | undefined;
@@ -88,7 +91,8 @@ export default class MapInteractionMode {
      * Determines whether a rectangle will be requested from the user rather than a set of pickedFeatures.
      */
     // this.drawRectangle = defaultValue(options.drawRectangle, false);
-
     this.onEnable = options.onEnable;
+
+    this.invisible = defaultValue(options.invisible, false);
   }
 }